remove logs, reorder helper functions, add csv dump
