+++ src/mappings/nabla/backstopPoolEventHandler.ts
@@ -18,7 +18,7 @@ import {
     getSwapPool,
 } from './creation'
 import { updateSwapPoolCoverageAndSupply } from './swapPoolEventHandler'
-import { addBackstopLP, removeBackstopLP } from '../handlePoints'
+import { addBackstopLP, removeBackstopLP } from '../points/handlePoints'
 
 export async function handleBackstopPoolEvent(
     ctx: EventHandlerContext,
@@ -202,7 +202,11 @@ export async function updateBackstopCoverageAndSupply(
     const contractHexAddress = ss58ToHex(backstopPool.id)
     const contract = new BackstopPoolContract(ctx, contractHexAddress)
     const poolTokenAddress = await contract.asset()
-    const poolTokenContract = new Erc20Contract(ctx, poolTokenAddress)
+    const poolTokenContract = new Erc20Contract(
+        ctx,
+        poolTokenAddress,
+        ctx.block.hash
+    )
     backstopPool.totalSupply = await contract.totalSupply()
     backstopPool.reserves = await poolTokenContract.balanceOf(
         contractHexAddress
