+++ src/mappings/nabla/creation.ts
@@ -78,9 +78,6 @@ export async function createBackstopPool(
     const ss58Address = hexToSs58(hexAddress)
     if (await getBackstopPool(ctx, ss58Address)) {
         // Unexpected error: backstop pool already exists at that address
-        console.log(
-            `createBackstopPool: backstop pool already exists at ${ss58Address}`
-        )
         return
     }
